services/vllm/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vllm-stack-autoconfig"
version = "0.1.0"
description = "Autoconfiguration, benchmark tuning, and docker compose generation for a local vLLM stack"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "Jinja2>=3.1",
  "requests>=2.31",
]

[project.scripts]
vllm-stack-manage = "vllm_stack.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
vllm_stack = ["templates/*.j2"]
